+++ tests/test_dataset.py
@@ -285,11 +285,11 @@ def test_mag_view_write_out_of_bounds() -> None:
     ds = Dataset(new_dataset_path)
     mag_view = ds.get_layer("color").get_mag("1")
 
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (24, 24, 24)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 24)
     mag_view.write(
         np.zeros((3, 1, 1, 48), dtype=np.uint8)
     )  # this is bigger than the bounding_box
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (24, 24, 48)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 48)
 
 
 def test_mag_view_write_out_of_bounds_mag2() -> None:
@@ -301,13 +301,13 @@ def test_mag_view_write_out_of_bounds_mag2() -> None:
     ds = Dataset(new_dataset_path)
     mag_view = ds.get_layer("color").get_or_add_mag("2-2-1")
 
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (0, 0, 0)
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (24, 24, 24)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (24, 24, 24)
     mag_view.write(
         np.zeros((3, 50, 1, 48), dtype=np.uint8), (10, 10, 10)
     )  # this is bigger than the bounding_box
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (0, 0, 0)
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (120, 24, 58)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (120, 24, 58)
 
 
 def test_update_new_bounding_box_offset() -> None:
@@ -316,21 +316,21 @@ def test_update_new_bounding_box_offset() -> None:
     ds = Dataset.create(TESTOUTPUT_DIR / "wk_dataset", scale=(1, 1, 1))
     mag = ds.add_layer("color", LayerCategories.COLOR_TYPE).add_mag("1")
 
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
 
     np.random.seed(1234)
     write_data = (np.random.rand(10, 10, 10) * 255).astype(np.uint8)
     mag.write(
         write_data, offset=(10, 10, 10)
     )  # the write method of MagDataset does always use the relative offset to (0, 0, 0)
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (10, 10, 10)
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (10, 10, 10)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (10, 10, 10)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (10, 10, 10)
 
     mag.write(
         write_data, offset=(5, 5, 20)
     )  # the write method of MagDataset does always use the relative offset to (0, 0, 0)
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (5, 5, 10)
-    assert tuple(ds.get_layer("color").get_bounding_box().size) == (15, 15, 20)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (5, 5, 10)
+    assert tuple(ds.get_layer("color").bounding_box.size) == (15, 15, 20)
 
 
 def test_write_multi_channel_uint8() -> None:
@@ -625,8 +625,8 @@ def test_view_write_without_open() -> None:
 
     ds = Dataset.create(ds_path, scale=(1, 1, 1))
     layer = ds.add_layer("color", LayerCategories.COLOR_TYPE)
-    layer.set_bounding_box(
-        offset=(0, 0, 0), size=(64, 64, 64)
+    layer.bounding_box = BoundingBox(
+        (0, 0, 0), (64, 64, 64)
     )  # This newly created dataset would otherwise have a "empty" bounding box
     mag = layer.add_mag("1")
 
@@ -689,35 +689,38 @@ def test_changing_layer_bounding_box() -> None:
     layer = ds.get_layer("color")
     mag = layer.get_mag("1")
 
-    bbox_size = ds.get_layer("color").get_bounding_box().size
+    bbox_size = ds.get_layer("color").bounding_box.size
     assert tuple(bbox_size) == (24, 24, 24)
     original_data = mag.read(size=bbox_size)
     assert original_data.shape == (3, 24, 24, 24)
 
-    layer.set_bounding_box_size((12, 12, 10))  # decrease bounding box
+    old_bbox = layer.bounding_box
+    old_bbox.size = np.array([12, 12, 10])
+    layer.bounding_box = old_bbox  # decrease bounding box
 
-    bbox_size = ds.get_layer("color").get_bounding_box().size
+    bbox_size = ds.get_layer("color").bounding_box.size
     assert tuple(bbox_size) == (12, 12, 10)
     less_data = mag.read(size=bbox_size)
     assert less_data.shape == (3, 12, 12, 10)
     assert np.array_equal(original_data[:, :12, :12, :10], less_data)
 
-    layer.set_bounding_box_size((36, 48, 60))  # increase the bounding box
+    old_bbox.size = np.array([36, 48, 60])
+    layer.bounding_box = old_bbox  # increase the bounding box
 
-    bbox_size = ds.get_layer("color").get_bounding_box().size
+    bbox_size = ds.get_layer("color").bounding_box.size
     assert tuple(bbox_size) == (36, 48, 60)
     more_data = mag.read(size=bbox_size)
     assert more_data.shape == (3, 36, 48, 60)
     assert np.array_equal(more_data[:, :24, :24, :24], original_data)
 
-    assert tuple(ds.get_layer("color").get_bounding_box().topleft) == (0, 0, 0)
+    assert tuple(ds.get_layer("color").bounding_box.topleft) == (0, 0, 0)
 
     # Move the offset from (0, 0, 0) to (10, 10, 0)
     # Note that the bottom right coordinate of the dataset is still at (24, 24, 24)
-    layer.set_bounding_box(offset=(10, 10, 0), size=(14, 14, 24))
+    layer.bounding_box = BoundingBox((10, 10, 0), (14, 14, 24))
 
-    new_bbox_offset = ds.get_layer("color").get_bounding_box().topleft
-    new_bbox_size = ds.get_layer("color").get_bounding_box().size
+    new_bbox_offset = ds.get_layer("color").bounding_box.topleft
+    new_bbox_size = ds.get_layer("color").bounding_box.size
     assert tuple(new_bbox_offset) == (10, 10, 0)
     assert tuple(new_bbox_size) == (14, 14, 24)
     # Note that even though the offset was changed (in the properties), the offset of 'mag.read()'
@@ -731,7 +734,7 @@ def test_changing_layer_bounding_box() -> None:
 
     # resetting the offset to (0, 0, 0)
     # Note that the size did not change. Therefore, the new bottom right is now at (14, 14, 24)
-    layer.set_bounding_box_offset((0, 0, 0))
+    layer.bounding_box = BoundingBox((0, 0, 0), new_bbox_size)
     new_data = mag.read()
     assert new_data.shape == (3, 14, 14, 24)
     assert np.array_equal(original_data[:, :14, :14, :], new_data)
@@ -1264,7 +1267,7 @@ def test_for_zipped_chunks() -> None:
         .get_or_add_mag("1", block_len=8, file_len=4)
     )
 
-    target_mag.layer.set_bounding_box(offset=(0, 0, 0), size=(256, 256, 256))
+    target_mag.layer.bounding_box = BoundingBox((0, 0, 0), (256, 256, 256))
     target_view = target_mag.get_view(size=(256, 256, 256))
 
     with get_executor_for_args(None) as executor:
@@ -1625,7 +1628,7 @@ def test_add_copy_layer(tmp_path: Path) -> None:
         cast(SegmentationLayer, ds.get_layer("segmentation")).largest_segment_id == 999
     )
     color_layer = ds.get_layer("color")
-    assert color_layer.get_bounding_box() == BoundingBox(
+    assert color_layer.bounding_box == BoundingBox(
         topleft=(10, 20, 30), size=(32, 64, 128)
     )
     assert color_layer.mags.keys() == original_color_layer.mags.keys()