+++ wkcuber/check_equality.py
@@ -116,7 +116,7 @@ def check_equality(
             source_mags == target_mags
         ), f"The mags between {source_path}/{layer_name} and {target_path}/{layer_name} are not equal: {source_mags} != {target_mags}"
 
-        official_bbox = wk_src_dataset.get_layer(layer_name).get_bounding_box()
+        official_bbox = wk_src_dataset.get_layer(layer_name).bounding_box
 
         for mag in mags:
             inferred_src_bbox = BoundingBox.from_auto(